// libs/feature/manager/offer/tracking/settings/view-info/data-access/src/lib/enums/invalid-traffic-forwarding.enum.ts
export enum InvalidTrafficForwardingEnum {
    MainDestinationUrl = 1,
    PreviewUrl,
    Offer,
    CustomTrafficbackUrl,
    GlobalTrafficbackUrl
}

export enum InvalidTrafficForwardingTranslateEnum {
    MainDestinationUrl = 'offers_page.tracking.fail_traffic_forwarding.main_destination_url',
    PreviewUrl = 'offers_page.tracking.fail_traffic_forwarding.preview_url',
    Offer = 'offers_page.tracking.fail_traffic_forwarding.offer',
    CustomTrafficbackUrl = 'offers_page.tracking.fail_traffic_forwarding.custom_trafficback_url',
    GlobalTrafficbackUrl = 'offers_page.tracking.fail_traffic_forwarding.global_trafficback_url'
}